Install WarpStream Agent/CLI with install script
This commit adds a new install script that allows users to easily install
the latest version of the WarpStream Agent/CLI on their machine. The
script detects the architecture and operating system of the machine
and downloads the appropriate version of the agent. It also updates the
user's profile to include the WarpStream executable in the PATH. After
installation, users can try the demo by running "warpstream demo".